Altium: Add support for dimension hight
ADDED: Parse the text position of a dimenson in a PcbDoc file.
Set the hight of the dimension accordingly.
This works when the dimmension is 90deg from the line between both
objects, everything else currently is not supported in KiCAD.
Dirrection can go the wrong way (complete opposite), even when
dimension is 90deg.